requirements.txt
streamlit
yfinance
plotly
pandas
redis
psycopg2-binary
streamlit-autorefresh
textblob
langchain
langchain-core
langchain-google-genai
langgraph
python-dotenv